requirements.txt
flask==3.0.2
google-generativeai==0.3.2
joblib==1.3.2
scikit-learn==1.4.1.post1
numpy==1.26.4
gunicorn==21.2.0 